calibrator: Use a skeleton symmetry specification
This symmetry specification is considered when calibrating the
locations of the joints with repect to the sensors. For instance,
the position of the hip w.r.t the sensor on the thigh and the
position of the knee w.r.t the same sensor imply the length of
the thigh. Obviously, the length of the left thigh is approximately
equal to the length of the right thigh. Up to know, such information
wasn't respected at all.